requirements.txt
pandas>=1.3.0
datasets>=2.0.0
huggingface-hub>=0.10.0
pyyaml>=6.0
requests>=2.28.0
openai>=1.0.0
nltk>=3.7
Pillow>=9.0
